add: useragent check

+++ app/index.js
@@ -4,6 +4,7 @@ const dbFile = process.env.DBFILE || "./data.sql";
 const db = new Database(dbFile);
 
 const short = require("short-uuid");
+const useragent = require("express-useragent");
 const express = require("express");
 const app = express();
 
@@ -12,6 +13,7 @@ const morgan = require("morgan");
 const url = process.env.URL || "http://localhost:3000";
 
 app.set("view engine", "ejs");
+app.use(useragent.express());
 app.use(morgan("combined"));
 app.use(express.urlencoded({ extended: true }));
 
@@ -25,24 +27,29 @@ app.get("/", (req, res) => {
   res.render("index", { form: true });
 });
 
-app.post("/geheim", (req, res) => {
+app.post("/", (req, res) => {
   const { text } = req.body;
   if (!text) return res.redirect("/");
 
   const uuid = short.generate().toLowerCase();
   db.prepare(`INSERT INTO ${tableName} VALUES (?,?)`).run(uuid, text);
 
-  res.render("index", { form: false, message: `${url}/geheim/${uuid}` });
+  res.render("index", { form: false, message: `${url}/${uuid}` });
 });
 
-app.get("/geheim/:id", (req, res) => {
-  const { id } = req.params;
-  if (id) {
-    const result = db.prepare("SELECT * FROM geheim WHERE uuid = ?").get(id);
-    if (!result) return res.redirect("/");
+app.get("/:id", (req, res) => {
+  const agent = req.useragent;
+  if (agent.isMobile || agent.isDesktop) {
+    const { id } = req.params;
+    if (id) {
+      const result = db.prepare("SELECT * FROM geheim WHERE uuid = ?").get(id);
+      if (!result) return res.redirect("/");
 
-    db.prepare("DELETE FROM geheim WHERE uuid = ?").run(id);
-    res.render("index", { form: false, message: result.message });
+      db.prepare("DELETE FROM geheim WHERE uuid = ?").run(id);
+      res.render("index", { form: false, message: result.message });
+    }
+  } else {
+    res.render("index", { form: true });
   }
 });
